Gaussian Elimination, Refactoring, RowVector Data Type
Moved all matrix operations into a separate class to reduce confusion in the Matrix class. The Matrix class should just be a "wrapper" data type to represent a matrix. Implimented another data type called RowVector to act as another "wrapper" data type to represent the row of a matrix when doing gaussian elimination. Implemented gaussian elimination in EliminationOperations class. Need to verify that it works as intended.